Use sqlite database as backend via gorm
This will now use an sqlite database for a backend, and return the title from there instead of wasting resources.
I've included time as a field because I should probably add some form of expiration or update process.

Also done a little tidying up with orders of various comments etc, and shifted the url handler off to its own function